Response and request model renamed New field added to entity Some related changes in unit test applied too

+++ AccountsApi/V1/Gateways/DynamoDbGateway.cs
@@ -34,6 +34,8 @@ namespace AccountsApi.V1.Gateways
 
         public async Task AddAsync(Account account)
         {
+            if(account==null)
+                throw new ArgumentNullException($"{nameof(account).ToString()} model shouldn't be null");
             await _dynamoDbContext.SaveAsync(account.ToDatabase()).ConfigureAwait(false);
         }
 
@@ -82,11 +84,15 @@ namespace AccountsApi.V1.Gateways
 
         public async Task RemoveAsync(Account account)
         {
+            if(account==null)
+                throw  new ArgumentNullException($"{nameof(account).ToString()} ModelStateExtension shouldn't be null");
             await _dynamoDbContext.DeleteAsync(account.ToDatabase()).ConfigureAwait(false);
         }
 
         public async Task UpdateAsync(Account account)
         {
+            if (account == null)
+                throw new ArgumentNullException($"{nameof(account).ToString()} ModelStateExtension shouldn't be null");
             await _dynamoDbContext.SaveAsync(account.ToDatabase()).ConfigureAwait(false);
         }
     }
